Adding all functionalities of adding and updating orders

+++ src/main/java/com/kartik/Ecommerce/model/Order.java
@@ -35,7 +35,7 @@ public class Order {
     @Embedded
     private PaymentDetails paymentDetails = new PaymentDetails();
 
-    private Double totalPrice;
+    private Integer totalPrice;
 
     private Integer totalDiscountedPrice;
     private Integer discount;
